refactor : separate class for person and relationship
