+++ src/SocialSim/Model/Model.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SocialSim.Model
+{
+    /// <summary>
+    /// Model class defines computation model for the simulation
+    /// It includes some static methods used in computation
+    /// </summary>
+    class Model
+    {
+    }
+}
